requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==10.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
unidecode==1.1.1 # https://pypi.org/project/Unidecode/

# Django
# ------------------------------------------------------------------------------
django==4.2.10  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==64.2.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-autoslug==1.9.8  # https://github.com/justinmayer/django-autoslug/

# Custom base packages
# ------------------------------------------------------------------------------
django-npm
anytree
python-dateutil
django-colorfield
schedule
django-crum==0.7.9
#django-storages==1.14.2
python-json-logger==2.0.7
django-ninja==1.3.0
inflect==7.2.1
pandas==2.0.3
reportlab==4.2.2
pyjwt==2.8.0

python-json-logger==2.0.7
sentry-sdk==2.14.0

redis==5.0.8
hiredis==3.0.0